backend/requirements.txt
# ============================================================================
# Face Kinship Verification - Complete Requirements
# ============================================================================
# RECOMMENDED: Python 3.11 (best compatibility and performance)
# Also works: Python 3.10
# NOT recommended: Python 3.12 (limited TensorFlow support)
# Installation: pip install -r requirements.txt
# ============================================================================

# Core Deep Learning Framework
tensorflow==2.16.2          # Latest stable, works with Python 3.9-3.12
keras==3.0.5                # Compatible with TensorFlow 2.16

# Image Processing
opencv-python==4.8.1.78     # Computer vision library
Pillow==10.1.0              # Image loading and manipulation

# Scientific Computing
numpy==1.26.4               # Arrays and numerical operations (compatible with TF 2.16)
scipy==1.11.4               # Scientific computing utilities

# Machine Learning Utilities
scikit-learn==1.3.2         # ML metrics and utilities

# Data Visualization (for training plots)
matplotlib==3.8.2           # Plotting library
seaborn==0.13.0             # Statistical visualizations

# Web Framework (for API)
Flask==3.0.0                # REST API framework
Flask-CORS==4.0.0           # Cross-origin resource sharing

# Optional but Recommended
tqdm==4.66.1                # Progress bars
h5py==3.10.0                # HDF5 file format (for model saving)

# ============================================================================
# Installation Notes:
# ============================================================================
# 1. Create virtual environment: python -m venv venv
# 2. Activate: venv\Scripts\activate (Windows) or source venv/bin/activate (Linux/Mac)
# 3. Upgrade pip: python -m pip install --upgrade pip
# 4. Install: pip install -r requirements.txt
# ============================================================================
